Update the code to handle the following case:
Imagine a test class import a src class but does not cover any lines
in the src class.  Assume that the test class only has a test method
testExecNothing() and this method is executed before another test
method testExecSomething() in another test class.  Assume the test
method testExecSomething() covers some lines in the src class.  In the
original implementation, testExecNothing() shares the same hitArray as
testExecSomething() and this may lead to incorrect computation of the
suspiciousness scores.  For example, if testExecSomething() fails,
then all lines covered by testExecSomething() will be treated as they
are executed twice as many times as they actually are because
testExecNothing()'s hitArray is updated by testExecSomething().